+++ src/node/rpc/node_call_types.h
@@ -186,15 +186,4 @@ namespace ccf
       size_t peak_allocated_heap_size = 0;
     };
   };
-
-  struct ObservedReconfigurationCommit
-  {
-    struct In
-    {
-      NodeId from;
-      kv::ReconfigurationId reconfiguration_id;
-    };
-
-    using Out = void;
-  };
 }
