+++ src/node/rpc/serialization.h
@@ -147,8 +147,4 @@ namespace ccf
 
   DECLARE_JSON_TYPE(UpdateResharing::In)
   DECLARE_JSON_REQUIRED_FIELDS(UpdateResharing::In, rid)
-
-  DECLARE_JSON_TYPE(ObservedReconfigurationCommit::In)
-  DECLARE_JSON_REQUIRED_FIELDS(
-    ObservedReconfigurationCommit::In, from, reconfiguration_id)
 }
